rewrote first-click logic
Logic before: select random point on board while conditions (non-bomb spot, not same spot) are not satisfied

The motivation to rewrite was due to logic bugs. Occasionally it would relocate bombs to points directly adjacent to the first click point. And this is inefficient as the method of solving this was through a while loop.

Logic after: create a "whitelist" of non-bomb points, then select a random point that is not whitelisted (with O(n) time wrt size of whitelist and O(1) time wrt points tested; the random point is GUARANTEED to be non-whitelisted)